LPD-96038 Move the insights view into the components directory
