+++ modules/dxp/apps/seo-studio/seo-studio-web/src/main/resources/META-INF/resources/js/components/insights_view/cell_renderers/ImpactCellRenderer.tsx
